requirements.txt
python-telegram-bot==21.5
requests==2.31.0
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1
pdfplumber==0.11.0
beautifulsoup4==4.12.3
faiss-cpu==1.8.0
sentence-transformers==2.7.0
numpy==1.26.4
python-dotenv==1.0.1
urllib3==2.1.0